feat: add validations and handle exceptions

+++ src/main/java/com/DevSprint/LibraryMS/controller/StaffController.java
@@ -15,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import com.DevSprint.LibraryMS.dto.StaffDTO;
+import com.DevSprint.LibraryMS.exception.StaffNotFoundException;
 import com.DevSprint.LibraryMS.service.StaffService;
 
 import lombok.RequiredArgsConstructor;
@@ -28,25 +29,61 @@ public class StaffController {
 
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> addStaffMember(@RequestBody StaffDTO staffDTO) {
+        if (staffDTO == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         staffService.addStaffMember(staffDTO);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @DeleteMapping
     public ResponseEntity<Void> deleteStaffMember(@RequestParam("staffId") String staffId) {
-        staffService.deleteStaffMember(staffId);
-        return ResponseEntity.noContent().build();
+        if (staffId == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        try {
+            staffService.deleteStaffMember(staffId);
+            return ResponseEntity.noContent().build();
+        } catch (StaffNotFoundException e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @PatchMapping(value = "/{staffId}", consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> updateStaffMember(@PathVariable String staffId, @RequestBody StaffDTO staffDTO) {
-        staffService.updateStaffMember(staffId, staffDTO);
-        return ResponseEntity.noContent().build();
+        if (staffId == null || staffDTO == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        try {
+            staffService.updateStaffMember(staffId, staffDTO);
+            return ResponseEntity.noContent().build();
+        } catch (StaffNotFoundException e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @GetMapping("/{staffId}")
     public ResponseEntity<StaffDTO> getStaffById(@PathVariable String staffId) {
-        return ResponseEntity.ok(staffService.getStaffById(staffId));
+        if (staffId == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        try {
+            return ResponseEntity.ok(staffService.getStaffById(staffId));
+        } catch (StaffNotFoundException e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @GetMapping
